+++ api/blueprints/drbCitation.py
@@ -1,3 +1,4 @@
+from socket import if_nametoindex
 from flask import Blueprint, request, current_app
 from ..db import DBClient
 from ..utils import APIUtils
@@ -58,9 +59,6 @@ def citationFetch(uuid):
         statusCode, 'citation', responseBody
     )
 
-# Why is 1900 the cutoff date for MLA citations?
-    #* Pre-1900: Use publication location in citation
-    #* Post-1900: Use publisher name in citation
 def mlaGenerator(citationWork):
     if citationWork:
         # Made title lowercase for consistency when finding this info in the database
@@ -88,62 +86,60 @@ def mlaGenerator(citationWork):
             
         else:
             return "MLA Citation not avaliable"
+        
     else:
         return "MLA Citation not avaliable"
 
-def mlaBibleGenerator(citationWork):
-    #Bible editions before 1900
-        #Specific version info of Bible currently not displayed in database
-    if citationWork.editions[0].publication_date < date(1900, 1, 1):
-        citationResponse = f"{citationWork.title}. Unknown version, \
-                            {citationWork.editions[0].publication_place}, \
-                            {citationWork.editions[0].publication_date.year}."
 
-    #Bible editions during and after 1900
-        #Specific version info of Bible currently not displayed in database
-    else:      
-        citationResponse = f"{citationWork.title}. Unknown version, \
-                            {citationWork.editions[0].publishers[0]['name']}, \
-                            {citationWork.editions[0].publication_date.year}."
-            
+#String that encompasses the publication info of a work
+    #* Why is 1900 the cutoff date for MLA citations?
+        #* Pre-1900: Use publication location in citation
+        #* Post-1900: Use publisher name in citation
+def generatePubString(edition):
+    pubString = ''
+    if edition.publication_date < date(1900, 1, 1):
+        pubString = f"{edition.publication_place}, \
+                    {edition.publication_date.year}"
+    else:
+        pubString = f"{edition.publishers[0]['name']}, {edition.publication_date.year}"
+
+    return pubString
+
+def mlaBibleGenerator(citationWork):
+
+    pubString = generatePubString(citationWork.editions[0])
+
+    #Specific version info of Bible currently not displayed in database
+    citationResponse = f"{citationWork.title}. Unknown version, {pubString}."
+      
     return ' '.join(citationResponse.split())
 
 def mlaGovGenerator(citationWork):
-    #Government Publications before 1900
-    if citationWork.editions[0].publication_date < date(1900, 1, 1):
-        
-        govName = citationWork.authors[0]['name']
+    govName = citationWork.authors[0]['name']
+    pubString = generatePubString(citationWork.editions[0])   
     
-        #Gov Pubs have government name and agency/division as the author
-            #* There's a period separating the Gov name and agency/division
-            #* Must replace with comma instead to match MLA rule
-        govNameAndAgency = govName.replace('.', ',', 1)
+    #Gov Pubs have government name and agency/division as the author
+        #* There's a period separating the Gov name and agency/division
+        #* Must replace with comma instead to match MLA rule
+    govNameAndAgency = govName.replace('.', ',', 1)
 
-        citationResponse = f"{govNameAndAgency}. \
-                    {citationWork.title}. \
-                    {citationWork.editions[0].publication_place}, \
-                    {citationWork.editions[0].publication_date.year}."
-
-    else:
-
-        govName = citationWork.authors[0]['name']
-        govNameAndAgency = govName.replace('.', ',', 1)
-
-        citationResponse = f"{govNameAndAgency}. \
-                            {citationWork.title}. \
-                            {citationWork.editions[0].publishers[0]['name']}, \
-                            {citationWork.editions[0].publication_date.year}."
-            
+    citationResponse = f"{govNameAndAgency}. {citationWork.title}. {pubString}."
+        
     return ' '.join(citationResponse.split())
 
-def mlaNoAuthorGenerator(citationWork):
+def nameSplit(name):
+    listName = name.split(', ')
+    lastName = listName[0]
+    firstName = listName[1]
+    return (firstName, lastName)
 
-    #Translated works/works prepared by editors
+def pubTransEditString(citationWork):
+
     if citationWork.contributors:
 
         i = 0
-        transName = ''
-        editorName = ''
+        transName = None
+        editorName = None
 
         while i < len(citationWork.contributors):
             if citationWork.contributors[i]['roles'] == 'translator':
@@ -153,615 +149,138 @@ def mlaNoAuthorGenerator(citationWork):
             i += 1
         
         if transName and editorName:
-            listTransName = transName.split(', ')
-            transLastName = listTransName[0]
-            transFirstName = listTransName[1]
-            listEditorName = editorName.split(', ')
-            editorLastName = listEditorName[0]
-            editorFirstName = listEditorName[1]
-
-            if citationWork.editions[0].publication_date < date(1900, 1, 1):
-
-                citationResponse = f"{citationWork.title}. \
-                                    Translated by {transFirstName} {transLastName}, \
-                                    edited by {editorFirstName} {editorLastName}, \
-                                    {citationWork.editions[0].publication_place}, \
-                                    {citationWork.editions[0].publication_date.year}."
-            else:
-
-                citationResponse = f"{citationWork.title}. \
-                                    Translated by {transFirstName} {transLastName}, \
-                                    edited by {editorFirstName} {editorLastName}, \
-                                    {citationWork.editions[0].publishers[0]['name']}, \
-                                    {citationWork.editions[0].publication_date.year}."
-
-            return ' '.join(citationResponse.split())
-                    
-        if transName:
-            listTransName = transName.split(', ')
-            transLastName = listTransName[0]
-            transFirstName = listTransName[1]
-            
-            if citationWork.editions[0].publication_date < date(1900, 1, 1):
-
-                citationResponse = f"{citationWork.title}. \
-                                    Translated by {transFirstName} {transLastName}, \
-                                    {citationWork.editions[0].publication_place}, \
-                                    {citationWork.editions[0].publication_date.year}."
-            else:
-
-                citationResponse = f"{citationWork.title}. \
-                                    Translated by {transFirstName} {transLastName}, \
-                                    {citationWork.editions[0].publishers[0]['name']}, \
-                                    {citationWork.editions[0].publication_date.year}."
-
-            return ' '.join(citationResponse.split())
+            transFirstName, transLastName = nameSplit(transName)
+            editorFirstName, editorLastName = nameSplit(editorName)
+            return f"Translated by {transFirstName} {transLastName}, \
+                    edited by {editorFirstName} {editorLastName}"
+        
+        elif transName:
+            transFirstName, transLastName = nameSplit(transName)
+            return f"Translated by {transFirstName} {transLastName}"
 
         elif editorName:
-            listEditorName = editorName.split(', ')
-            editorLastName = listEditorName[0]
-            editorFirstName = listEditorName[1]
+            editorFirstName, editorLastName = nameSplit(editorName)
+            return f"edited by {editorFirstName} {editorLastName}"
+    
+    return None
 
-            if citationWork.editions[0].publication_date < date(1900, 1, 1):
+def editionStatement(editStatement):
+    return f"{editStatement}" if editStatement else None
 
-                citationResponse = f"{citationWork.title}, \
-                                    edited by {editorFirstName} {editorLastName}, \
-                                    {citationWork.editions[0].publication_place}, \
-                                    {citationWork.editions[0].publication_date.year}."
-            else:
 
-                citationResponse = f"{citationWork.title}, \
-                                    edited by {editorFirstName} {editorLastName}, \
-                                    {citationWork.editions[0].publishers[0]['name']}, \
-                                    {citationWork.editions[0].publication_date.year}."
+def mlaNoAuthorGenerator(citationWork):
+    transEditString = pubTransEditString(citationWork)
+    pubString = generatePubString(citationWork.editions[0])
+    editStatement = editionStatement(citationWork.editions[0].edition_statement)
 
-            return ' '.join(citationResponse.split())
-
+    # return translated and/or editor work
+    if transEditString:
+        if transEditString.split(' ')[0] == 'edited':
+            citationResponse = f"{citationWork.title}, {transEditString}, {pubString}."
         else:
-            return mlaNoAuthorExtraRulesGenerator(citationWork)
+            citationResponse = f"{citationWork.title}. {transEditString}, {pubString}."
+
+    # return multiple edition work
+    elif editStatement:
+        citationResponse = f"{citationWork.title}. {editStatement}, {pubString}."
+
+     # return single edition work
     else:
-        return mlaNoAuthorExtraRulesGenerator(citationWork)
-            
-#Works with no authors that are not translated or prepared by editors
-def mlaNoAuthorExtraRulesGenerator(citationWork):
-    #Single edition books with no author before 1900
-    if citationWork.editions[0].publication_date < date(1900, 1, 1) and \
-        citationWork.editions[0].edition_statement is None:
+        citationResponse = f"{citationWork.title}. {pubString}."
 
-        citationResponse = f"{citationWork.title}. {citationWork.editions[0].publication_place}, \
-                            {citationWork.editions[0].publication_date.year}."
+    return ' '.join(citationResponse.split())
 
-        return ' '.join(citationResponse.split())
-                
-    #Multiple edition books with no author before 1900
-    elif citationWork.editions[0].publication_date < date(1900, 1, 1) and \
-        citationWork.editions[0].edition_statement is not None:
-
-        citationResponse = f"{citationWork.title}. {citationWork.edition_statement}, \
-                            {citationWork.editions[0].publication_place}, \
-                            {citationWork.editions[0].publication_date.year}."
-
-        return ' '.join(citationResponse.split())
-
-    #Single edition books with no author during and after 1900
-    elif citationWork.editions[0].publication_date >= date(1900, 1, 1) and \
-        citationWork.editions[0].edition_statement is None:
-
-        citationResponse = f"{citationWork.title}. {citationWork.editions[0].publishers[0]['name']}, \
-                            {citationWork.editions[0].publication_date.year}." 
-
-        return ' '.join(citationResponse.split())
-
-    #Multiple edition books with no author during and after 1900
-    elif citationWork.editions[0].publication_date >= date(1900, 1, 1) and \
-        citationWork.editions[0].edition_statement is not None:
-
-        citationResponse = f"{citationWork.title}. {citationWork.edition_statement}, \
-                            {citationWork.editions[0].publishers[0]['name']}, \
-                            {citationWork.editions[0].publication_date.year}." 
-
-        return ' '.join(citationResponse.split())
-
-    else:
-        return "MLA Citation rule not implemented"
         
 def mlaOneAuthorGenerator(citationWork):
 
-    #Translated works/works prepared by editors 
-    if citationWork.contributors:
+    authName = citationWork.authors[0]['name']
+    transEditString = pubTransEditString(citationWork)
+    pubString = generatePubString(citationWork.editions[0])
+    editStatement = editionStatement(citationWork.editions[0].edition_statement)
 
-        i = 0
-        transName = ''
-        editorName = ''
-
-        while i < len(citationWork.contributors):
-            if citationWork.contributors[i]['roles'] == 'translator':
-                transName = citationWork.contributors[i]['name']
-            if citationWork.contributors[i]['roles'] == 'editor':
-                editorName = citationWork.contributors[i]['name']
-            i += 1
-        
-        if transName and editorName:
-            listTransName = transName.split(', ')
-            transLastName = listTransName[0]
-            transFirstName = listTransName[1]
-            listEditorName = editorName.split(', ')
-            editorLastName = listEditorName[0]
-            editorFirstName = listEditorName[1]
-            authName = citationWork.authors[0]['name']
-
-            if citationWork.editions[0].publication_date < date(1900, 1, 1):
-
-                citationResponse = f"{authName}. {citationWork.title}. \
-                                    Translated by {transFirstName} {transLastName}, \
-                                    edited by {editorFirstName} {editorLastName}, \
-                                    {citationWork.editions[0].publication_place}, \
-                                    {citationWork.editions[0].publication_date.year}."
-            else:
-
-                citationResponse = f"{authName}. {citationWork.title}. \
-                                    Translated by {transFirstName} {transLastName}, \
-                                    edited by {editorFirstName} {editorLastName}, \
-                                    {citationWork.editions[0].publishers[0]['name']}, \
-                                    {citationWork.editions[0].publication_date.year}."
-
-            return ' '.join(citationResponse.split())
-                    
-        if transName:
-            listTransName = transName.split(', ')
-            transLastName = listTransName[0]
-            transFirstName = listTransName[1]
-            authName = citationWork.authors[0]['name']
-
-            if citationWork.editions[0].publication_date < date(1900, 1, 1):
-
-                citationResponse = f"{authName}. {citationWork.title}. \
-                                    Translated by {transFirstName} {transLastName}, \
-                                    {citationWork.editions[0].publication_place}, \
-                                    {citationWork.editions[0].publication_date.year}."
-            else:
-
-                citationResponse = f"{authName}. {citationWork.title}. \
-                                    Translated by {transFirstName} {transLastName}, \
-                                    {citationWork.editions[0].publishers[0]['name']}, \
-                                    {citationWork.editions[0].publication_date.year}."
-
-            return ' '.join(citationResponse.split())
-
-        elif editorName:
-            listEditorName = editorName.split(', ')
-            editorLastName = listEditorName[0]
-            editorFirstName = listEditorName[1]
-            authName = citationWork.authors[0]['name']
-
-            if citationWork.editions[0].publication_date < date(1900, 1, 1):
-
-                citationResponse = f"{authName}. {citationWork.title}, \
-                                    edited by {editorFirstName} {editorLastName}, \
-                                    {citationWork.editions[0].publication_place}, \
-                                    {citationWork.editions[0].publication_date.year}."
-            
-            else:
-
-                citationResponse = f"{authName}. {citationWork.title}, \
-                                    edited by {editorFirstName} {editorLastName}, \
-                                    {citationWork.editions[0].publishers[0]['name']}, \
-                                    {citationWork.editions[0].publication_date.year}."
-
-            return ' '.join(citationResponse.split())
-        
+    # return translated and/or editor work
+    if transEditString:
+        # Editor work has comma after title instead of period
+        if transEditString.split(' ')[0] == 'edited':
+            citationResponse = f"{authName}. {citationWork.title}, {transEditString}, {pubString}."
         else:
-            return mlaOneAuthorExtraRulesGenerator(citationWork)
+            citationResponse = f"{authName}. {citationWork.title}. {transEditString}, {pubString}."
+
+    # return multiple edition work
+    elif editStatement:
+        # Book by corporate author or organization
+        if authName == citationWork.editions[0].publishers[0]['name']:
+            citationResponse = f"{citationWork.title}. {editStatement}, {pubString}."
+
+        else:
+            citationResponse = f"{authName}. {citationWork.title}. {editStatement}, {pubString}."
+
+    # return single edition work
     else:
-        return mlaOneAuthorExtraRulesGenerator(citationWork)
-
-#Works with one author that are not translated or prepared by editors
-def mlaOneAuthorExtraRulesGenerator(citationWork):
-    #Single edition books with one author before 1900
-    if citationWork.editions[0].publication_date < date(1900, 1, 1) and \
-        citationWork.editions[0].edition_statement is None:
-
-        authName = citationWork.authors[0]['name']
-
-        #Book by corporate author or organization
+        # Book by corporate author or organization
         if authName == citationWork.editions[0].publishers[0]['name']:
-            citationResponse = f"{citationWork.title}. \
-                                {citationWork.editions[0].publishers[0]['name']}, \
-                                {citationWork.editions[0].publication_place}, \
-                                {citationWork.editions[0].publication_date.year}."
+            citationResponse = f"{citationWork.title}. {pubString}."
 
-            return ' '.join(citationResponse.split())
+        else:
+            citationResponse = f"{authName}. {citationWork.title}. {pubString}."
 
-        else:   
-            citationResponse = f"{authName}. {citationWork.title}. \
-                                {citationWork.editions[0].publishers[0]['name']}, \
-                                {citationWork.editions[0].publication_place} \
-                                {citationWork.editions[0].publication_date.year}."
-
-            return ' '.join(citationResponse.split())
-
-    #Multiple edition books with one author before 1900
-    elif citationWork.editions[0].publication_date < date(1900, 1, 1) and \
-        citationWork.editions[0].edition_statement is not None:
-
-        authName = citationWork.authors[0]['name']
-
-        #Book by corporate author or organization
-        if authName == citationWork.editions[0].publishers[0]['name']:
-            citationResponse = f"{citationWork.title}. {citationWork.edition_statement}, \
-                                {citationWork.editions[0].publishers[0]['name']}, \
-                                {citationWork.editions[0].publication_place}, \
-                                {citationWork.editions[0].publication_date.year}."
-            return ' '.join(citationResponse.split())
-
-        else:   
-            citationResponse = f"{authName}. {citationWork.title}. {citationWork.edition_statement}, \
-                                {citationWork.editions[0].publishers[0]['name']}, \
-                                {citationWork.editions[0].publication_place} \
-                                {citationWork.editions[0].publication_date.year}."
-
-            return ' '.join(citationResponse.split())
-    
-        
-    #Single edition books with one author during and after 1900
-    elif citationWork.editions[0].publication_date >= date(1900, 1, 1) and \
-        citationWork.editions[0].edition_statement is None:
-        
-        authName = citationWork.authors[0]['name']
-
-        #Book by corporate author or organization
-        if authName == citationWork.editions[0].publishers[0]['name']:
-            citationResponse = f"{citationWork.title}. \
-                                {citationWork.editions[0].publishers[0]['name']}, \
-                                {citationWork.editions[0].publication_date.year}."
-            return ' '.join(citationResponse.split())
-
-        else:   
-            citationResponse = f"{authName}. {citationWork.title}. \
-                                {citationWork.editions[0].publishers[0]['name']}, \
-                                {citationWork.editions[0].publication_date.year}."
-            return ' '.join(citationResponse.split())
-
-    #Multiple Edition books with one author during and after 1900
-    elif citationWork.editions[0].publication_date >= date(1900, 1, 1) and \
-        citationWork.editions[0].edition_statement is not None:
-
-        authName = citationWork.authors[0]['name']
-
-        #Book by corporate author or organization
-        if authName == citationWork.editions[0].publishers[0]['name']:
-            citationResponse = f"{citationWork.title}. {citationWork.edition_statement} \
-                                {citationWork.editions[0].publishers[0]['name']}, \
-                                {citationWork.editions[0].publication_date.year}."
-            return ' '.join(citationResponse.split())
-
-        else:   
-            citationResponse = f"{authName}. {citationWork.title}. {citationWork.edition_statement} \
-                                {citationWork.editions[0].publishers[0]['name']}, \
-                                {citationWork.editions[0].publication_date.year}."
-            return ' '.join(citationResponse.split())
-
-    else:
-        return "MLA Citation rule not implemented"
+    return ' '.join(citationResponse.split())
 
 def mlaTwoAuthorsGenerator(citationWork):
 
-    #Translated works/works prepared by editors
-    if citationWork.contributors:
+    authName1 = citationWork.authors[0]['name']
+    authName2 = citationWork.authors[1]['name']
+    listAuthName2 = authName2.split(', ')
+    auth2LastName = listAuthName2[0]
+    auth2FirstName = listAuthName2[1]
+    
+    transEditString = pubTransEditString(citationWork)
+    pubString = generatePubString(citationWork.editions[0])
+    editStatement = editionStatement(citationWork.editions[0].edition_statement)
 
-        i = 0
-        transName = ''
-        editorName = ''
-
-        while i < len(citationWork.contributors):
-            if citationWork.contributors[i]['roles'] == 'translator':
-                transName = citationWork.contributors[i]['name']
-            if citationWork.contributors[i]['roles'] == 'editor':
-                editorName = citationWork.contributors[i]['name']
-            i += 1
-        
-        if transName and editorName:
-            listTransName = transName.split(', ')
-            transLastName = listTransName[0]
-            transFirstName = listTransName[1]
-            listEditorName = editorName.split(', ')
-            editorLastName = listEditorName[0]
-            editorFirstName = listEditorName[1]
-
-            authName1 = citationWork.authors[0]['name']
-            authName2 = citationWork.authors[1]['name']
-            listAuthName2 = authName2.split(', ')
-            auth2LastName = listAuthName2[0]
-            auth2FirstName = listAuthName2[1]
-
-            if citationWork.editions[0].publication_date < date(1900, 1, 1):
-
-                citationResponse = f"{authName1}, and {auth2FirstName} {auth2LastName}. \
-                                    {citationWork.title}. \
-                                    Translated by {transFirstName} {transLastName}, \
-                                    edited by {editorFirstName} {editorLastName}, \
-                                    {citationWork.editions[0].publication_place}, \
-                                    {citationWork.editions[0].publication_date.year}."
-            else:
-
-                citationResponse = f"{authName1}, and {auth2FirstName} {auth2LastName}. \
-                                    {citationWork.title}. \
-                                    Translated by {transFirstName} {transLastName}, \
-                                    edited by {editorFirstName} {editorLastName}, \
-                                    {citationWork.editions[0].publishers[0]['name']}, \
-                                    {citationWork.editions[0].publication_date.year}."
-
-            return ' '.join(citationResponse.split())
-                    
-        if transName:
-            listTransName = transName.split(', ')
-            transLastName = listTransName[0]
-            transFirstName = listTransName[1]
-
-            authName1 = citationWork.authors[0]['name']
-            authName2 = citationWork.authors[1]['name']
-            listAuthName = authName2.split(', ')
-            auth2LastName = listAuthName[0]
-            auth2FirstName = listAuthName[1]
-
-            if citationWork.editions[0].publication_date < date(1900, 1, 1):
-
-                citationResponse = f"{authName1}, and {auth2FirstName} {auth2LastName}. \
-                                    {citationWork.title}. \
-                                    Translated by {transFirstName} {transLastName}, \
-                                    {citationWork.editions[0].publication_place}, \
-                                    {citationWork.editions[0].publication_date.year}."
-            else:
-
-                citationResponse = f"{authName1}, and {auth2FirstName} {auth2LastName}. \
-                                    {citationWork.title}. \
-                                    Translated by {transFirstName} {transLastName}, \
-                                    {citationWork.editions[0].publishers[0]['name']}, \
-                                    {citationWork.editions[0].publication_date.year}."
-
-            return ' '.join(citationResponse.split())
-
-        elif editorName:
-            listEditorName = editorName.split(', ')
-            editorLastName = listEditorName[0]
-            editorFirstName = listEditorName[1]
-
-            authName1 = citationWork.authors[0]['name']
-            authName2 = citationWork.authors[1]['name']
-            listAuthName2 = authName2.split(', ')
-            auth2LastName = listAuthName2[0]
-            auth2FirstName = listAuthName2[1]
-            
-            if citationWork.editions[0].publication_date < date(1900, 1, 1):
-
-                citationResponse = f"{authName1}, and {auth2FirstName} {auth2LastName}. \
-                                    {citationWork.title}, \
-                                    edited by {editorFirstName} {editorLastName}, \
-                                    {citationWork.editions[0].publication_place}, \
-                                    {citationWork.editions[0].publication_date.year}."
-            else:
-
-                citationResponse = f"{authName1}, and {auth2FirstName} {auth2LastName}. \
-                                    {citationWork.title}, \
-                                    edited by {editorFirstName} {editorLastName}, \
-                                    {citationWork.editions[0].publishers[0]['name']}, \
-                                    {citationWork.editions[0].publication_date.year}."
-
-            return ' '.join(citationResponse.split())
+    # return translated and/or editor work
+    if transEditString:
+        # Editor work has comma after title instead of period
+        if transEditString.split(' ')[0] == 'edited':
+            citationResponse = f"{authName1}, and {auth2FirstName} {auth2LastName}. \
+                                 {citationWork.title}, {transEditString}, {pubString}."
 
         else:
-            return mlaTwoAuthorsExtraRulesGenerator(citationWork)
+            citationResponse = f"{authName1}, and {auth2FirstName} {auth2LastName}. \
+                                 {citationWork.title}. {transEditString}, {pubString}."
+    # return multiple edition work
+    elif editStatement:
+        citationResponse = f"{authName1}, and {auth2FirstName} {auth2LastName}. \
+                             {citationWork.title}. {editStatement}, {pubString}."
+
+    # return single edition work
     else:
-        return mlaTwoAuthorsExtraRulesGenerator(citationWork)
-
-#Works with two authors that are not translated or prepared by editors
-def mlaTwoAuthorsExtraRulesGenerator(citationWork):
-    #Single edition books with two authors before 1900
-    if citationWork.editions[0].publication_date < date(1900, 1, 1) and \
-        citationWork.editions[0].edition_statement is None:
-
-        authName1 = citationWork.authors[0]['name']
-        authName2 = citationWork.authors[1]['name']
-        listAuthName2 = authName2.split(', ')
-        auth2LastName = listAuthName2[0]
-        auth2FirstName = listAuthName2[1]
-
         citationResponse = f"{authName1}, and {auth2FirstName} {auth2LastName}. \
-                            {citationWork.title}. {citationWork.editions[0].publication_place}, \
-                            {citationWork.editions[0].publication_date.year}."
+                             {citationWork.title}. {pubString}."
 
-        return ' '.join(citationResponse.split())
-
-    #Multiple edition books with two authors before 1900
-    elif citationWork.editions[0].publication_date < date(1900, 1, 1) and \
-        citationWork.editions[0].edition_statement is not None:
-
-        authName1 = citationWork.authors[0]['name']
-        authName2 = citationWork.authors[1]['name']
-        listAuthName = authName2.split(', ')
-        auth2LastName = listAuthName[0]
-        auth2FirstName = listAuthName[1]
-
-        citationResponse = f"{authName1}, and {auth2FirstName} {auth2LastName}. \
-                            {citationWork.title}. {citationWork.edition_statement} \
-                            {citationWork.editions[0].publication_place}, \
-                            {citationWork.editions[0].publication_date.year}."
-
-        return ' '.join(citationResponse.split())
-            
-    #Single Edition books with two authors during and after 1900
-    elif citationWork.editions[0].publication_date >= date(1900, 1, 1) and \
-        citationWork.editions[0].edition_statement is None:
-
-        authName1 = citationWork.authors[0]['name']
-        authName2 = citationWork.authors[1]['name']
-        listAuthName2 = authName2.split(', ')
-        auth2LastName = listAuthName2[0]
-        auth2FirstName = listAuthName2[1]
-                
-        citationResponse = f"{authName1}, and {auth2FirstName} {auth2LastName}. \
-                            {citationWork.title}. {citationWork.editions[0].publishers[0]['name']}, \
-                            {citationWork.editions[0].publication_date.year}."
-            
-        return ' '.join(citationResponse.split())
-
-    #Multiple Edition books with two authors during and after 1900
-    elif citationWork.editions[0].publication_date >= date(1900, 1, 1) and \
-        citationWork.editions[0].edition_statement is not None:
-
-        authName1 = citationWork.authors[0]['name']
-        authName2 = citationWork.authors[1]['name']
-        listAuthName2 = authName2.split(', ')
-        auth2LastName = listAuthName2[0]
-        auth2FirstName = listAuthName2[1]
-                
-        citationResponse = f"{authName1}, and {auth2FirstName} {auth2LastName}. \
-                            {citationWork.title}. {citationWork.editions[0].edition_statement}, \
-                            {citationWork.editions[0].publishers[0]['name']}, \
-                            {citationWork.editions[0].publication_date.year}."
-            
-        return ' '.join(citationResponse.split())
-
-    else:
-        return "MLA Citation rule not implemented"
+    return ' '.join(citationResponse.split())
 
 def mlaThreeAuthorsGenerator(citationWork):
 
-    #Translated works/works prepared by editors before 1900
-    if citationWork.contributors:
+    authName = citationWork.authors[0]['name']
+    transEditString = pubTransEditString(citationWork)
+    pubString = generatePubString(citationWork.editions[0])
+    editStatement = editionStatement(citationWork.editions[0].edition_statement)
 
-        i = 0
-        transName = ''
-        editorName = ''
-
-        while i < len(citationWork.contributors):
-            if citationWork.contributors[i]['roles'] == 'translator':
-                transName = citationWork.contributors[i]['name']
-            if citationWork.contributors[i]['roles'] == 'editor':
-                editorName = citationWork.contributors[i]['name']
-            i += 1
-        
-        if transName and editorName:
-            listTransName = transName.split(', ')
-            transLastName = listTransName[0]
-            transFirstName = listTransName[1]
-            listEditorName = editorName.split(', ')
-            editorLastName = listEditorName[0]
-            editorFirstName = listEditorName[1]
-            authName = citationWork.authors[0]['name']
-
-            if citationWork.editions[0].publication_date < date(1900, 1, 1):
-
-                citationResponse = f"{authName}, et al. {citationWork.title}. \
-                                    translated by {transFirstName} {transLastName}, \
-                                    edited by {editorFirstName} {editorLastName}, \
-                                    {citationWork.editions[0].publication_place}, \
-                                    {citationWork.editions[0].publication_date.year}."
-            else:
-                
-                citationResponse = f"{authName}, et al. {citationWork.title}. \
-                                    translated by {transFirstName} {transLastName}, \
-                                    edited by {editorFirstName} {editorLastName}, \
-                                    {citationWork.editions[0].publishers[0]['name']}, \
-                                    {citationWork.editions[0].publication_date.year}."
-
-            return ' '.join(citationResponse.split())
-                    
-        if transName:
-            listTransName = transName.split(', ')
-            transLastName = listTransName[0]
-            transFirstName = listTransName[1]
-            authName = citationWork.authors[0]['name']
-
-            if citationWork.editions[0].publication_date < date(1900, 1, 1):
-
-                citationResponse = f"{authName}, et al. {citationWork.title}. \
-                                    translated by {transFirstName} {transLastName}, \
-                                    {citationWork.editions[0].publication_place}, \
-                                    {citationWork.editions[0].publication_date.year}."
-            else:
-
-                citationResponse = f"{authName}, et al. {citationWork.title}. \
-                                    translated by {transFirstName} {transLastName}, \
-                                    {citationWork.editions[0].publishers[0]['name']}, \
-                                    {citationWork.editions[0].publication_date.year}."
-
-            return ' '.join(citationResponse.split())
-
-        elif editorName:
-            listEditorName = editorName.split(', ')
-            editorLastName = listEditorName[0]
-            editorFirstName = listEditorName[1]
-            authName = citationWork.authors[0]['name']
-
-            if citationWork.editions[0].publication_date < date(1900, 1, 1):
-                
-                citationResponse = f"{authName}, et al. {citationWork.title}, \
-                                    edited by {editorFirstName} {editorLastName}, \
-                                    {citationWork.editions[0].publication_place}, \
-                                    {citationWork.editions[0].publication_date.year}."
-            else:
-
-                citationResponse = f"{authName}, et al. {citationWork.title}, \
-                                    edited by {editorFirstName} {editorLastName}, \
-                                    {citationWork.editions[0].publishers[0]['name']}, \
-                                    {citationWork.editions[0].publication_date.year}."
-
-            return ' '.join(citationResponse.split())
+    # return translated and/or editor work
+    if transEditString:
+        # Editor work has comma after title instead of period
+        if transEditString.split(' ')[0] == 'edited':
+            citationResponse = f"{authName}, et al. {citationWork.title}, {transEditString}, {pubString}."
         else:
-            mlaThreeAuthorsExtraRulesGenerator(citationWork)
+            citationResponse = f"{authName}, et al. {citationWork.title}. {transEditString}, {pubString}."
+
+    # return multiple edition work
+    elif editStatement:
+        citationResponse = f"{authName}, et al. {citationWork.title}. {editStatement}, {pubString}."
+
+    # return single edition work
     else:
-        mlaThreeAuthorsExtraRulesGenerator(citationWork)
+        citationResponse = f"{authName}, et al. {citationWork.title}. {pubString}."
 
-#Works with three authors that are not translated or prepared by editors
-def mlaThreeAuthorsExtraRulesGenerator(citationWork):
-    #Single Edition books with three authors or more before 1900
-    if citationWork.editions[0].publication_date < date(1900, 1, 1) and \
-        citationWork.editions[0].edition_statement is None:
-            
-        authName = citationWork.authors[0]['name']
-
-        citationResponse = f"{authName}, et al. {citationWork.title}. \
-                {citationWork.editions[0].publication_place}, \
-                {citationWork.editions[0].publication_date.year}."
-
-        return ' '.join(citationResponse.split())
-        
-    #Multiple edition books with three authors or more before 1900
-    elif citationWork.editions[0].publication_date < date(1900, 1, 1) and \
-        citationWork.editions[0].edition_statement is not None:
-            
-        authName = citationWork.authors[0]['name']
-
-        citationResponse = f"{authName}, et al. {citationWork.title}. \
-                            {citationWork.editions[0].edition_statement} \
-                            {citationWork.editions[0].publication_place}, \
-                            {citationWork.editions[0].publication_date.year}."
-
-        return ' '.join(citationResponse.split())
-        
-    #Single Edition books with three authors during and after 1900
-    elif citationWork.editions[0].publication_date >= date(1900, 1, 1) and \
-        citationWork.editions[0].edition_statement is None:
-
-        authName = citationWork.authors[0]['name']
-
-        citationResponse = f"{authName}, et al. {citationWork.title}. \
-                {citationWork.editions[0].publishers[0]['name']}, {citationWork.editions[0].publication_date}."
-
-        return ' '.join(citationResponse.split())
-
-    #Multiple edition books with three authors during and after 1900
-    elif citationWork.editions[0].publication_date >= date(1900, 1, 1) and \
-        citationWork.editions[0].edition_statement is not None:
-
-        authName = citationWork.authors[0]['name']
-
-        citationResponse = f"{authName}, et al. {citationWork.title}. {citationWork.editions[0].edition_statement}, \
-                            {citationWork.editions[0].publishers[0]['name']}, \
-                            {citationWork.editions[0].publication_date.year}."
-
-        return ' '.join(citationResponse.split())
-
-    else:
-        return "MLA Citation rule not implemented"
+    return ' '.join(citationResponse.split())